data/afcl_workflows/BWA/Functions/GCP/bwaSplit/requirements.txt
apache-libcloud==3.3.1
boto3==1.18.33
botocore==1.21.33
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
idna==3.2
jmespath==0.10.0
multimethod==1.5
pycparser==2.20
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
numpy